// src/shared/configuration/configuration.enum.ts
export enum Configuration {
  APP_PORT = "APP_PORT",
  APP_HOST = "APP_HOST",
  NODE_ENV = "NODE_ENV",
  DB_TYPE = "DB_TYPE",
  DB_HOST = "DB_HOST",
  DB_PORT = "DB_PORT",
  DB_NAME = "DB_NAME",
  DB_USERNAME = "DB_USERNAME",
  DB_PASSWORD = "DB_PASSWORD",
  SESSION_SECRET = "SESSION_SECRET",
  SESSION_EXPIRE_TIME_MS = "SESSION_EXPIRE_TIME_MS",
  REDIS_HOST = "REDIS_HOST",
  REDIS_PORT = "REDIS_PORT"
}